Update packages before install Python dependencies
On circle CI "build & push" job, we started having errors on the build
when trying to install Python dependencies.
The errors are related with missing remote .deb packages. And this is
happening as we do not "apt-get update" before "apt-get install"
Adding the update before the install calls fixes it.