Break CircleCI Test into Lint, Unit and Integr.
This will allow the Lint checks, the unit tests and the Integration
tests to run in parallel.
It makes each job slower due to the required setup but will achieve
a faster Test workflow overall.

Extract shared CircleCI steps into aliases

Now that we have different jobs for running Lint checks, Unit tests
and Integration tests, multiple steps are shared between these jobs.

This make our jobs setup a little bit too verbose and hard to
understand and maintain.

By extracting those steps to aliases shared between the jobs, we can
focus our attention on the key bits of the jobs that make a difference,
instead of having to navigate through all the boilerplating.

We also identify and remove steps that are not needed in some of the
jobs, making them faster.