Add text box for the respondent's police notified question for the 'yes' answer.
This was to keep the existing functionality for police notified 'no' answer except allow a textbox to be used for when the answer is 'yes' as-well.

Update respondent form to validate the text box if the answer is 'yes' or 'no' (true or false)

Remove :police_notified from the clear_details method as the same text box is used for a yes or no answer and both require it.

Create respondent javascript file to update the text box label depending on the 'yes' or 'no' answer for police notified. Also update the error messages.

Update the view to always show the textbox for police notified, which will then be hidden by the javascript. This is because the requirement is to have the text box hidden as default and appear for yes and no answers on police_notified.

Update tests