Remove JavaScript (Ajax requests) on the Statement of case page.
Use the upload button for JS enabled and disabled browsers. This improves accessibility alerts as the ajax requests that were on the page meant were on the 'choose a file' button which overrode any custom aria-alerts set.

Remove multifile uploads and replace them with single file uploads. This makes it easier to test the file and return a success or failure.

Add the following messages to be used with aria-alerts/accessibility alerts:

successfully_deleted

error_message

successful_upload

Remove PDF Files Table as it was not being called or used anywhere

Update tests and translation files