Remove mailer tests in the request tests
This is because in rails 6 the test queue adapter is overriding sidekiq in request tests. The ResendLinkRequestMailer and FeedbackMailer is already tested seperately in mailer tests.

Tried to use alternative solutions such as disabling the queue adapter but that would cause errors elsewhere e.g. in a email monitor service test. ((ActiveJob::Base.descendants << ActiveJob::Base).each { |a| a.disable_test_adapter })