Modify CircleCI to auto deploy master merges
What
deploy_uat will be an optional step on non-master branches but,
when code is merged to master, it will build and deploy to UAT
then wait for approval to push to staging and prod environments.

Why
This will ensure that UAT master is always up to date and surface
any issues around the deploy step before it gets to the staging
environment.  e.g. a dependabot merge that breaks deploy could
be merged to master but not actually deployed until some days
later when code is changed and we may not be sure if the code or
gem update broke the deploy

How
Update the circle config file to:-
* make hold_uat and deploy_uat ignore the master branch
* add an auto_deploy_uat workflow step that will _only_ run on
master branch