create new delete_uat_release script
This is based on ruby and is designed to delete all out
of date uat tiller instances.

It builds two lists, one of open pull requests on git hub
and one of the existing tiller instances.  It compares them
and, if the PR matching the instance name no longer exists,
deletes, and purges, the tiller instance