Dev-update: Add Guard properly
A Guardfile was added in a previous PR, this tidies that file up
and adds the requisite gems.

Guard is used to monitor code changes and run the affected tests
e.g. if you change the `app/models/applicant.rb` file while guard
is running, it will run the tests in `spec/models/applicant_spec.rb`
and then it will then run rubocop against the file too.

To run all tests, simply hit enter in the terminal running guard
and it will execute rspec and rubocop against all files.

There is an issue with the guard-cucumber plugin at the moment, so
I have excluded until the issues are resolved