Update feature tests to categorise bank transactions
Update feature tests to categorise bank transactions into the transaction types on the legal aid application model

If the user clicks save and continue and there are transaction types with no bank transactions then it will re-render that index page.

Otherwise, it will allow the user to resume with the application process.