Introduce gather route to wait for bank_transactions_analyser before means_summary
Update bank_transactions_analyser_job to a worker. This makes it a sidekiq job, which gives features included in the sidekiq-status
gem we already use such as: check the status of a job and is it :complete, :working , etc.

Create a gather route for the means summary controller with a view of its own. This is because bank_transactions_analyser would perform_later and then
once it got its results it would change the state of the application. The state machine was not waiting and instead redirected users to the check_passported_answers
page instead of the /means_summaries page. As found in the flow.

To combat this, reuse worker-waiter javascript, which keeps calling the gather route in means summaries until the status of the worker is :complete, which it then redirects
to the means summaries show page.

Update the flow so that it goes from the other_assets to gather_means_transactions_results, whilst bank_transactions_analyser is working in the background. It will
show a loading gif whilst it is pending.

Move the change state / state event legal_aid_application.complete_bank_transaction_analysis! into the controller after the bank_transactions_analyser worker completes.

Add a column into the legal aid application, gathering_transactions_jid to replace sessions on the gather page. It checks to see if that JobID/worker is still pending.
Make it a view helper, which is used in the gather view and worker-waiter javascript file.

Update locales

Update feature tests