Add a loop to wait_for_ajax and refresh page
This is because the controller keeps rendering the /gather page whilst the
sidekiq job is still running in the background. Once complete then it'll allow the use to move forward.

The loop is set at 10 times and the wait_for_ajax is 2 seconds , so 20 seconds in total. Will refactor this if it works.