Fix screen reader bug with visually hidden field still being announced
Fix voiceover bug with using class 'govuk-visually-hidden' because the screen reader reads whatever is inside it
aloud even if it's hidden. Instead use class 'govuk-radios__conditional--hidden' for this, which hides it from the screen reader too if
it's not showing on the page.

e.g. visually hidden (visually hides) but doesn't hide from the screenreaders/the ear.

// conditional--hidden will hide from voiceover/screen readers compared to visually hidden
+  const govuk_hidden_class = 'govuk-radios__conditional--hidden'