Enforce ordering of CCMS histories in specs
When running postgres12, obtain_applicant_reference_service_spec has
failures due to SubmissionHistory records not being created in the
expected order.

This commit adds an order clause to ensure that the correct record is
always used for comparison.

It also includes a minor tidyup of the code including renaming some
things, and moving the 'applicant does not exist on CCMS' context within
the 'operation successful' context.