Add a bank transaction presenter
This is used to build a single row for the CSV export
of bank transaction data. Each entry in the CELLS collection
requires a matching `transaction_XXXX` method that will be
called as the row is generated.
I chose this method instead of meta_programming for ease of
use, this could be refactored in the future if needed