Add new rubocop cops
To fix this warning:

The following cops were added to RuboCop, but are not configured. Please
set Enabled to either `true` or `false` in your `.rubocop.yml` file:
 - Performance/AncestorsInclude (1.7)
 - Performance/BigDecimalWithNumericArgument (1.7)
 - Performance/RedundantSortBlock (1.7)
 - Performance/RedundantStringChars (1.7)
 - Performance/ReverseFirst (1.7)
 - Performance/SortReverse (1.7)
 - Performance/Squeeze (1.7)
 - Performance/StringInclude (1.7)
For more information: https://docs.rubocop.org/rubocop/versioning.html

All new cops are added and enabled. `rubocop -a` was run to fix failures