Add change client details link to use_ccms
When a provider receives a negative DWP check result on the passported
journey, they do not have the option to change the applicant's details.
They do have this option on the non-passported journey.

This commit moves the code which displays the applicant's details,
including the link to change them, to a new partial which is now used on
both pages. It also refactors the providers translations to move
wordings to the new partial, and adds a feature test for this scenario.