AP-1490 Pass benefit label to CFE instead of name
When Apply sends state benefit data to CFE it is currently including the
metadata name of the benefit. CFE uses this value to classify the
benefit and include it in/exclude it from the means calculation.

However CFE is expecting to receive the metadata label of the benefit.
This means that none of the benefits received by CFE can be correctly
classified: they are all being classified as 'other' and included in the
means calculation.

This commit amends the CreateStateBenefitsService so that the label is
passed to CFE instead of the name, and fixes some broken specs
accordingly.