AP-1490 Disregarded benefits not being sent to CFE
When the state_benefits payload is created to be sent to CFE, it is
only including transactions marked as 'benefits' not 'excluded_benefits'.

This commit adds excluded_benefits to the select logic in
create_state_benefits_service so that all benefit transactions are
passed to CFE. It also fixes a bug in the bank_transactions factory
which was masking this problem.