AP-1505 Add checks to legal aid application policy to ensure provider has right permissions to manage applications
After the DWP check, the application is either "passported" or "non-passported".  Providers have passported and/or passported
permissions.  We only allow passported applications to be managed by providers with passported permissions, and ditto for
non-passported.

We dont' want to carry out this check before the DWP check, so the ProviderBaseController has been subclassed to
PreDWPCheckProviderBaseController, and all controllers up to the DWP check inherit from this, rather than directly
from the ProviderBaseController.

Authorizable has been changed to pass in an AuthorizationContext which is just the provider and the controller.
In this way, the pundit policy can determine whether or not to carry out the passported/non-passported check.