AP-1461 Refactor state machine into two
The LegalAidApplicationStateMachine has become too big and is doing two
things - managing both the passported and non-passported flows.

This commit breaks the state machine into three parts, a BaseStateMachine
which includes states used by both flows, and two subclasses,
PassportedStateMachine and NonPassportedStateMachine, to handle the
relevant states for those flows.

It adds functionality to the LegalAidApplication to create and manage
the state machine.

The majority of changes in the commit are to specs and feature tests
to ensure that test  data is created properly using this new pattern.
Traits have been added  and amended in the LegalAidApplication factory
to support this.