AP-1557 Display correct means result when manual_check_required is set
When the feature to require manual checking of applications is enabled,
incorrect means results are displayed.

The same text is always displayed regardless of the result, and refers
to restrictions on assets, even if restrictions do not apply. This text
should only be displayed if restrictions do apply; in all other cases
the correct result text should be displayed.

To fix this, this commit amends the logic in the `overview` method of
the CFE::BaseResult class so that `manual_check_required` is only
returned when restrictions apply. In all other scenarios the correct
result will be returned and the correct text displayed on screen.