AP-1378 Fix colour contrast on check_benefits
On /check_benefits, there are two places where, when a link has focus,
white text is displayed on a yellow background. This contrast is not
accessible.

There are two different solutions to fix these:

1) remove inline styling from the details__summary in the
_how_we_checked partial which was setting the text colour to white, and
replace it with CSS. This styling has lower precedence and allows the
inherited black colour to be used as expected when the element has focus.

2) add a CSS class to style links as black, and add this in the relevant
place.

In addition to fixing the two identified issues in the ticket, it also
fixes them in two additional places - on the use_ccms screens after
a failed benefit check (on the passported journey) and if consent to
online banking is not given.