Update Circle CI
This does two things to the circle merge & deploy process

It makes the deploy to master UAT optional
* for the duration of the DAC tetsing we should avoid merging code
unless it is an accessibility update required for them to complete tests

* it runs the production hold notification as a stub, this allows us to
deploy to production without waiting for the message to appear in slack,
sometimes for bugfixes, we are monitoring closely and the addition 30
second delay can be excruciating.  It also mirrors the flow in CFE, but
I forgot to replicate it back when I made the original changes!