Add a screenreader alert when proceeding types search result is returned
Add a span tag with class attribute 'govuk-visually-hidden' and a role attribute with alert.

Update the span tag with javascript when proceeding types result is returned. This causes the role alert to alert
a screen reader that a change has been made and reads out the message to tab out of the search box when results appear below the search box.