Add accessibility message for safari browser voiceover usage
Add aria-label to the input field to describe that it's a search box and to tab out to see results. This is for accessibility users.

Accomadate for safari browsers accessiblity users that use voiceover. This is because Safari Broswer does not read out role="alert" alerts
if the focus is inside an input text (Something i've noticed, please feel free to re-test this)