Fix accessibility issues
Add the aria-describedby id to the select_all_that_apply field as it broke WAVE testing of aria tags for the savings and investments page and other assets page.

Fix duplication of error messages under the select_all_that_apply hint because it's already in the error summary above the title. This is only when checkboxes are checked and nested values cause an error on submission. It still shows an error underneath it for when no checkboxes were selected validation fails

Test the above. However, keep an eye on https://github.com/rspec/rspec-expectations/pull/1168
which is an rspec matcher coming soon that allows to test how many times something has been seen in a response body.