AP-1409 Simplify negative DWP result page
The DWP result page for applicants that do not receive benefits contains
too much information and needs to be simplified.

This commit makes changes to the _check_benefits_result partial and
providers translations to display the new design, and ensure that the
page continues to be displayed correctly for other results.

A new partial, _applicant_details, is added which is reused by
both _check_benefits_result and _how_we_checked partials.