AP-1606 Display residual balance flag on means report
If a CFE result is received that has a residual_balance remark (ie the
applicant's capital is greater than the lower threshold and their
current account balance is greater than 0), it needs to be flagged for
caseworker review on the means report.

This is mostly done automatically by existing functionality. This flag
is slightly different to existing remarks as it does not need a
category. This commit amends the remarks model and caseworker review
partial to check the type of remark, and only populates the category if
appropriate. It also adds a translation to display the correct wording.