AP-1643 Clear TrueLayer errors
After receiving an error authenticating with TrueLayer, if a user
tries again and successfully authenticates, the same error is still
displayed.

This commit fixes this problem by setting session[:worker_id] = nil in
accounts_controller after the error has been handled. This removes the
worker from the session and when the second attempt is made a new worker
is created. This prevents the error from being replayed.