AP-1641 Improve display of TrueLayer errors
If errors are received when authenticating with a bank through TrueLayer
an ugly error string is displayed on the page. This commit improves that
by extracting the description of the error and displaying that, or a
default error if no description can be found.

It does several things:

- Add methods to the accounts_controller to parse the error and extract
the relevant description or provide a default error.
- Amends the error generated by the api_client to make it easier to
parse.
- Removes an error being set by consents_controller if authentication
failed. This was having the effect of making it look as if the error
was caused by the user not giving consent, when this was not the case.
- Wraps the error in a page_template so that a back button is generated
to allow the user to navigate away from the error.