AP-1638 Add running_balance to bank_transaction
Caseworkers need to know the running balance of an account at any
given transaction. This data is passed to us by TrueLayer. This commit
adds a running_balance field to the bank_transaction model, and amends
the import_transactions_service to extract the data from the TrueLayer
response and persist it to the database.