Fix back button behaviour in merits screens
There is inconsistent behaviour when using the back links in the
merits screens. It often takes the user back much further than required,
into the means screens, and results in a lot of clicking to get back to
the correct place.

This commit hardcodes the correct back link path into each view in the
merits flow.