AP-1612 Make screen readers ignore null table cells
In two places there are tables that contain deliberately empty cells -
the provider's applications list and the transaction selection screens.

These are read out by screen readers which is confusing for users. The
null cells cannot be easily removed as they are used to rearrange the
layout of the pages on small screens.

To prevent the cells being announced by screen readers, mark them as
aria-hidden="true".