AP-1644 Prevent duplicate bank providers
If a user authenticates twice with the same bank through TrueLayer, a
duplicate bank_provider record is created resulting in duplicate
bank_account, bank_account_holder, and bank_transaction records.

This duplication occurs because ImportProviderService checks for
existing bank_providers on credentials_id, however credentials_id is
always unique and so no duplicate will ever be found.

This commit changes this functionality so that it instead checks for
duplicates based on true_layer_provider_id, which will be consistent for
all bank details retrieved from a given bank.