Fix start button event handler
While working on AP-1616 I noticed that the event handler in
start_button.js, used to make start buttons respond to the space bar,
is firing an error on all pages where there is no 'start' button.

Add a check for nulls to prevent this happening.

Also the default space bar behaviour of scrolling is firing when a user
uses the keyboard to activate the button, which looks a bit strange.

Add preventDefault to stop this happening.