AP-1616 Make continue buttons accessible
On application_confirmations/show.html.erb, the 'Back to your application'
link has been styled as a button but does not have a role of button
assigned or a keyboard event handler to allow keyboard only users to
activate the element with the space bar.

Add a role of 'button' to the link and create an event handler to allow
a user to activate the buttin with the space bar.