Fix flaky postcode test
The OS postcode checker previously didn't return any addresses for
SW1H 9AJ, so it is used in specs as a test for non-existent addresses.

However 102 Petty France is now (correctly) returned for that postcode
which will cause tests to fail next time VCR cassettes are recorded.

This commit changes SW1H 9AJ to XX1X 1XX which is vaild postcode that is
unlikely to have any addresses associated with it the near future.