AP-1648 Add CCMS applicant if no applicant ref returned
ObtainApplicantReferenceService sends an Add Applicant request to CCMS
if there are no matching applicants in the CCMS database - ie the
record_count field in the response is zero.

However a recent production incident highlighted that it's possible for
an applicant to exist in the CCMS database but be disabled, in which
case a non-zero record_count is returned, but no ClientReferenceNumber
is included. This causes failures later in the submission process

This commit amends ObtainApplicantReferenceService so that an Add
Applicant request is sent in any situation where a ClientReferenceNumber
cannot be derived from the CCMS response.