Reduce calls to Geckoboard
We're seeing lots of Sentry alerts due to exceeding the number of
permitted calls to Geckoboard.

This seems to be linked to the ProviderDataJob, which is triggered by
a save to the LegalAidApplication model. The call is guarded by an if
statement that always resolves to true:

'if proc { |laa| laa.state.eql?(:assessment_submitted) }'

This returns the proc, which is truthy, so a Geckoboard request is sent
every time the LegalAidApplication is saved. This commit corrects this
so that the request is only made if the state is :assessment_submitted.