AP-1625 More descriptive links for screenreaders
Add a helper method to create links with text that is partially hidden.
This allows us to keep to our current design but also add extra context
for screenreader users - for example a link can be displayed as 'Change'
but read out as 'Change postcode'.

Use this helper for links on the

* about_the_financial_assessment
* address_selection
* income_summary
* outgoings_summary
* and limititations

screens.