AP-1625 Make remove dependant link more accessible
This wasn't included in AP-1625 but is similar to the issues identified
there so I'm taking the opportunity to fix it now.

On has_other_dependants/show.html.erb, it is possible to have multiple
Remove links (one for each depandant). Each link would be read as
'Remove' which might be confusing for screenreader users.

This commit adds the defendant name to the link, hidden so that it is
not displayed on screen but is read out by screenreaders.

AccessibleLinkHelper can't be used here as these are buttons-as-links,
not actual links. This appears to be the only place where this pattern
is used so it's not worth creating a new helper yet.