Add extra state transition
We're seeing production errors due to a valid state transition.

It's possible for a provider to process a non-passported case that
uses delegated functions, get to the non_passported_client_instructions
screen, then hit back twice.

This currently throws the error "Event 'applicant_details_checked'
cannot transition from 'delegated_functions_used'".

This commit adds the delegated_functions_used state to the
applicant_details_checked event so that this is allowed.