AP-1756 Add Content Security Policy
To improve security and remove XSS vulnerabilities, add and configure
the secure_headers gem in order to add content-security-policy headers
to responses.

As a result of this, a change to inline javascript is needed to prevent
it being blocked by the policy. This is done by referencing a nonce
generated by the gem.