AP-1713 Store page history in redis
Add a service class to write/read page history records to/from redis,
based on a key based on the session_id.

Amend Backable module and ConfirmOfficesController to use the new
service class instead of using the session to store the page history.

Fix specs and feature tests to ensure that a consistent session id is
available throughout a test, to allow data to be be written/read
correctly.