Update the procceding_type form
This is the partial that renders the proceeding type search results

It removes non alphanumeric characters from the 'select and continue'
button to make the text less stilted.
e.g. Extend, variation or discharge - Part IV would be read as
Extend <pause> variation or discharge <pause> part 4

It also marks the visual display as aria-hidden, this prevents the
text being read twice