BUGFIX: update pundit
When attempting to delete an old claim a user was shown an authoriasation error

Because it was an old claim, it had been marked as `use_ccms` because the benefit
check failed.  This also meant that the application failed the `passported?` check
and theerfore the provider needed non-passported permissions to delete the claim.

This checks the state_machine_proxy and returns the correct permissions. It may
be worth nothing that this is a bit of a stricking plaster fix and that other
legacy claims may be in non-passported states due to the migration and that we
should endeavour to assess the liklihood of this recurring.