Add architecture diagram to README
This commit adds the architecture diagram for Apply to the README.

The CI status badge is also moved to the top along with some new
headings for technical information and dependencies to improve the
structure.

The description seemed like it was for back when this was the API
half of Apply, so it has been updated to a more accurate description
of what the system really does.