AP-1751 Fix HTML Injection
In several places, text entered by the user is echoed back to the screen
on check your answers pages using the 'simple_format' method. This
renders HTML to the screen and allows for HTML injection.

This commit removes 'simple_format' from anywhere it is used to display
user-entered text. It remains in places where it is used to display
text from translation files.

It also fixes a small layout error caused by the removal of
'simple_format' by increasing padding in one partial.