Fix layout error on merits CYA
There is an error in views/shared/check_answers/_merits.html.erb which
causes the layout to be slightly incorrect - the section break is
appearing between "Have the police been notified?" and the associated
text, instead of after the text.

This fixes the partial so that the page displays correctly - text is
always required for this question, regardless of whether yes or no is
selected, so this removes the logic around displaying the section break.