BUGFIX: remove names from cuke step
When the delete function was added there were two anchor tags on the row
and this step was amended to look for the anchor tag containing the
applicants name.  This breaks when faker generates a name with an
apostrophe e.g. O`Malley.

This check has been removed and it now looks for the anchor tag in the
first cell of the row instead