Update feedback controller to include application id AP-1727 Update the feedback controller to include the application id. Attain the application id from different sources depending on the scenario e.g. When the provider clicks on feedback/signout when on an application then Use the PageHistoryService / page_history to get the application id.
Update saml session controller to leave a session called 'signed_out' with value true when the user signs out. This is then used in the feedback controller
to ultimately let the user know that they can close their browser tab. As opposed to go back to their applications if they gave feedback whilst signed in.

Give the feedback mailer the application id as a param.

Rename citizen to 'Applicant' which is used in the feedback mailer.