AP-1676 Display all errors in Welsh
Error pages were not displaying in Welsh when the locale=cy param was
set. This appears to be because the error page is rendered before
around_action :switch_locale executes in the controller, resulting in
the default locale (en) always being used.

Instead of relying on switch_locale to set the correct language, code
that renders errors now explicitly sets the locale pased on the
params[:locale] value.

In order to also display errors that occur outside of the application in
Welsh (eg /unknown/path, as opposed to /citizens/unknown/path), the
catch-all error route in config/routes now directs to a new
page_not_found method in the errors_controller.