AP-1676 Persist locale after truelayer callback
After the callback from TrueLayer, the locale parameter is lost and all
subsequent pages will default to English. To allow a user to continue
their journey in Welsh, store the locale value in the session at the
step before starting the TrueLayer process and set the locale from the
session when returned from TrueLayer.