Update ProviderDetailsService
We have had errors raised where an attempt to add a user failed. We
believe that it was a copy and paste from an outlook email or word
document the text `brand new‑user` was parsed as `BRAND%20NEW\u2011USER`

This adds a test for those circumstances and displays an error on the
admin pages requesting the text be re-typed from scratch.

If the user did it on the portal I would expect the portal to handle the
error on it's own