tech-debt: update circle flow
In some of the work to ignore webhint and turn uat deploys
off and on again some of the jobs got orphaned.

It was possible for the cucumber tests to fail and still push to live
It is still possible for master uat deploy to fail but as the helm
work is progressing I will leave that one for now!