AP-1673 No language switcher when fetching bank data
The language switcher is currently displayed on
/citizens/gather_transactions while bank transactions are being
downloaded or confirmation is displayed.

If a user clicks to change language on this screen it restarts the
TrueLayer download process and the user has to wait for the whirling to
stop.

To prevent this, add validation to the page_template_helper so that the
language switcher is not displayed on this screen.