AP-1673 Add language-switcher component
As per designs, add a small component to allow applicants to switch
language between English and Welsh. This will be present at the
top-right of the page, aligned with the Back link (if present).

This commit adds styling, translations, and a helper method to generate
the links and modifies application.html.erb to display the component
when the Welsh language setting is enabled.