Remove webhint from Rake
Webhint runs as part of the rake test suite. It usually fails. We've
removed it from the CircleCI pipeline so it also makes sense to remove
it here, replacing it with Cucumber so that the feature tests still run.

Webhint can still be run locally by using the webhint:generate_reports
rake task.