AP-1673 Prevent provider pages appearing in Welsh
If a user completes an applicant journey in Welsh, closes the browser
tab, and then starts a provider journey, some of the provider screens
appear in Welsh (or with missing translations).

This is mostly an issue during development/testing and is unlikely to
happen in production although could be possible if a provider and
applicant share a device.

To prevent it occurring, call update_locale on the start_controller to
force the app into English when a provider first logs in.