AP-1673 Refactor language switcher logic
Using the journey_type method to identify which pages should have the
language switcher displayed wasn't working accurately enough,
particularly on the feedback, request a new link, and assessment already
completed pages.

This commit reverts the changes made accommodate that
functionality and instead adds a session variable which is set when
the applicant accesses the service and can be used to identify any pages
they have visited.