Enable zeitwerk autoloader
Rails 6 now uses the `zeitwerk` autoloader by default. When Apply was
migrated to Rails 6 this caused problems so we reverted to the `classic`
autoloader.

This commit enables `zeitwerk` and fixes issues to allow it to function
correctly. These all relate to class names - `zeitwerk` requires strict
camelcase and does not play nicely with acronyms. To resolve this I have
added inflections for common acronyms and ensured consistency in the way
those acronyms in the code.