AP-1673 Enable language-switcher component
Adds code to allow the language-switcher component to be displayed on
relevant applicant screens. This is based on the journey_type variable,
which is good for most of the applicant journey.

In order for this to function on other screens that applicant can visit
that do not fall into the 'citizens' domain (eg, error pages, feedback,
contacts, privacy policy) it has been necessary to split out some of
the functionality from Flowable into a new modules
JourneyTypeIdentifiable and LocaleSwitchable and apply them to relevant
controllers.