BUGFIX: error on legal_aid_application_spec.rb:278
This was caused by the transaction type factory picking a value from
TransactionType::NAMES.values.flatten.sample

When this was excluded_benefits the handling was just different
enough to cause the test to fail.  Because it was randomised within
the factory re-running rspec with the same seed would not provide the
same type on a subsequent run!