AP-1804 Add MimeMagic gem to validate file types
There is a security vulnerability in the way we check for valid
statement of case file types using the file's content_type, which can
be forged in HTTP requests.

Instead add the MimeMagic gem, which can examine the content of the
file itself, and use this in the StatementOfCaseForm to check that the
file is of the correct type.

This causes a small issue as MimeMagic does not identify empty files in
the way we expect, so this commit also amends the logic in the
original_file_valid method to check for empty files before it checks for
valid file types.