Remove redis initializer
Changes to the redis gem were causing deprecation warnings to be
constantly displayed by sidekiq. We added a redis initializer to
suppress those warnings.

The issue with redis has subsequently been fixed however our initializer
is now causing the deprecation warning to be displayed when running eg
rspec or cucumber from the command line.

Removing the initializer prevents the deprecation warning from being
displayed on the command line and does not reintroduce the warning to
sidekiq.