AP-1496 Accessible Statement of Case delete links
The links used to delete uploaded statements of case are all read out as
'Delete' by screen readers. This can be improved by including the
attachment name for the screen reader.

These links are actually `button-as-link`s. To add the desired
functionality replicate the existing `link_to_with_hidden_suffix` helper
method for buttons, creating a `button_to_with_hidden_suffix` method
that appends the attachment name with a `govuk-visually-hidden` class
so that it is read out by screen readers but does not appear onscreen.