AP-1888 Show error summaries on additional accounts
Error summaries are not being displayed on the two additional accounts
screens when errors exist. This is because the errors are produced by
validation in the controller, not the form or model and so the
page template helper is not aware of them.

This commit adds an error partial which is rendered by each view if an
error is present.

It also adds an extra error translation for the /new page as the existing
error 'Select yes if you have accounts with other banks'
(which is displayed on both screens) did not make sense. 'Select yes if
you have online access to your accounts with other banks' is clearer.