Capture exceptions during entity attributes generation
During the ccms submission we generate entitiy attributes

This commit captures the error and sends a message to Sentry via Raven explaining the config
. This allows for better debugging.

After rescuing the error , it reraises it. It'll be rescued again later. This rescue duplication is so that we can check
what the config looked like in the attribute generation.

Test this