AP-1859 Tidy up bank account holder code
Noticed during work on AP-1859, reviewing placeholder CCMS values. We're
currently hardcoding bank account holder as 'Client Sole' in two places.

This commit changes this so that we only hardcode it in the BankAccount
model and the attribute generator now reads that value.

* Fixes an error in the BankAccount model where 'ClientSole' should be
'Client Sole'.

* Renames bank_account_holders to bank_account_holder_type in various
places to remove possible confusion with the BankAccountHolder model
used elsewhere, which relates to the actual person who holds the account
as opposed to the type of account (eg sole/joint).