Fix omniauth CSRF vulnerability
Omniauth 2.0.0 fixes the longstanding CSRF vulnerability. To get it to
work properly with Apply, versions of the omniauth-google-oauth2,
omniauth-oauth2 and omniauth-rails_csrf_protection gems need to be
bumped to.

omniauth_csrf_vulnerability_spec is no longer necessary and will
fail as a result of these changes, so the file has been removed.