Update proceeding types controller and add service object
Update the proceeding types controller #create route which wasn't being used instead the update route was.
The create route now adds a proceeding type with a scope limitation of substantive. If it fails, it adds an error to select a proceeding type.

Add a service class to take the logic of adding a proceeding with a scope limitation and encapsulating it. This makes it easier to read.

Update the request tests to test the new flow when the feature flag is turned on

Add a service object unit test