AP-1962 Rename overriden ccms_submission method
The legal_aid_application model overrides the default ccms_submission
method, causing it to create a new submission when it is called (if one
doesn't already exist). This is undesirable, so this commit renames the
method, calls it in the appropriate place, and fixes some tests which
break as a result.