AP-1964 Post cash payments to CFE
Update the CFE SubmissionManager to include a call to the new
CreateCashTransactionsService as part of the non-passported request.

At the moment this is a separate call as it is only done if the
allow_cash_payment setting is enabled. When cash payments are live it
can be added to the NON_PASSPORTED_SERVICES array, and the
'allow_cash_payment setting is true' context removed from the spec.

Also updates the CFE submission state machine.