Undo #in_time_zone for DateTime objects
This is because this ruby file is not requiring activesupport/all or active support timezones etc.
This means that the DateTime object (Date) is not extended and doesn't have in_time_zone method.

Set this to use dates without fear of timezone changes. This is because the app is using England/London timezones e.g. the date of the image in ecr is UK.