AP-2014 Refactor date params
Following the switch to the DfE formbuilder we can use a Rails form
helper to refine the params we permit. Instead of permitting each part
of the date (_1i, _2i, _3i) individually, we can simply permit the whole
date.

This commit amends the used_delegated_functions controller and
date_client_told_incident controller.

In addition to amending the permitted params it is necessary to change
the way the convert_date_params method is called - instead of being
called before permitting the params it must be called after and
requires a small refactor as a result. This has a minor knock-on effect
on two other controllers (applicants and applicant_details) which
handle dates but can't be changed in the same way.